database: rename dbh to database_handler
since it is more meaningful
