+++ eduu/database/__init__.py
@@ -1,4 +1,4 @@
 # SPDX-License-Identifier: MIT
 # Copyright (c) 2018-2021 Amano Team
 
-from .dbh import *
+from .database_handler import *
